+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.x branch
 ## 1.8 branch
 ### 1.8.2
+* `TestTrait::assertException()` is deprecated and will be removed in a later release;
 * `TestTrait::expectAssertionFailed()` is deprecated and will be removed in a later release.
 
 ### 1.8.1